Roles, claims, crud claims

+++ IdentityApp/Pages/Identity/Admin/Dashboard.cshtml.cs
@@ -1,14 +1,20 @@
+using IdentityApp.Core.Configurations;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using IdentityApp.Core.Extensions;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Options;
 
 namespace IdentityApp.Pages.Identity.Admin {
 
-    public class DashboardModel : AdminPageModel {
-
-        public DashboardModel(UserManager<IdentityUser> userMgr)
-            => UserManager = userMgr;
+    public class DashboardModel : AdminPageModel
+    {
+        private readonly IdentityInitializeOptions _options;
+        public DashboardModel(UserManager<IdentityUser> userMgr, IOptions<IdentityInitializeOptions> options)
+        {
+            UserManager = userMgr;
+            _options = options.Value;
+        }
 
         public UserManager<IdentityUser> UserManager { get; set; }
 
@@ -35,8 +41,12 @@ namespace IdentityApp.Pages.Identity.Admin {
 
         public async Task<IActionResult> OnPostAsync() {
             foreach (IdentityUser existingUser in UserManager.Users.ToList()) {
-                IdentityResult result = await UserManager.DeleteAsync(existingUser);
-                result.Process(ModelState);
+                if (emails.Contains(existingUser.Email) ||
+                    !await UserManager.IsInRoleAsync(existingUser, _options.Role))
+                {
+                    IdentityResult result = await UserManager.DeleteAsync(existingUser);
+                    result.Process(ModelState);
+                }
             }
             foreach (string email in emails) {
                 IdentityUser userObject = new IdentityUser {
